Update C-Chess knight strikes back.cpp

+++ Lesson 8/C-Chess knight strikes back.cpp
@@ -25,9 +25,9 @@ int main() {
         cout << 8;
     } else {
         ll sm = 0;
-        vector <int> a = {1, 1, 1, 1, 0, 1, 1, 0, 1, 0};
+        vector <ll> a = {1, 1, 1, 1, 0, 1, 1, 0, 1, 0};
         for (int i = 1; i < n; i++){
-            vector <int> b (10, 0);
+            vector <ll> b (10, 0);
             b[0] = a[5] + a[7];
             b[1] = a[6] + a[8];
             b[2] = a[3] + a[7];
